Created a reflection.
I copied the sprite flipped it and put it bellow the other sprite. I then flipped it so it should give a reflection.
I think it's a bit much at the moment so I will probably edit it soon but you can see the idea of it.

Also I currently don't reflect the border I should probably do this.